moved all GUI builder functions into their invidivudal files
