+++ Misc Tools/Limits Version Control Tool/misc.py
@@ -1,7 +1,17 @@
 "Misc functions for the limits_tool_gui"
 
+import os.path
+from PyQt5.QtWidgets import QPushButton, QFrame
 from PyQt5.QtGui import QDesktopServices
 from PyQt5.QtCore import QUrl
+
+
+def get_user_directory(directory):
+    home = os.path.expanduser("~")
+    storage_dir = os.path.join(home, directory)
+    if not os.path.exists(storage_dir):
+        os.makedirs(storage_dir)
+    return storage_dir
 
 
 def validate_all_conditions(self):
@@ -48,3 +58,28 @@ def update_master_button_layout(self):
 def open_sheets_master():
     url = QUrl("https://docs.google.com/spreadsheets/d/15rRk5JAMWXBGiKTly4aP0cUuFE1qECZe01tNESSKXBo/edit?usp=sharing")
     QDesktopServices.openUrl(url)
+
+
+def create_separator(self):
+    line= QFrame(); line.setFrameShape(QFrame.HLine)
+    line.setFrameShadow(QFrame.Sunken)
+    return line
+
+
+def add_sheet_master_btn(self):
+    "Add the 'Open Sheets Master' button to the GUI"
+    self.btn_master= QPushButton("Open Sheets Master")
+    self.btn_master.setEnabled(True)
+    self.btn_master.clicked.connect(lambda: open_sheets_master())
+    self.layout.addWidget(self.btn_master)
+    update_master_button_layout(self)
+
+
+def add_exit_btn(self):
+    "Add the exit button to the GUI"
+    self.layout.addStretch()
+    self.btn_exit= QPushButton("Exit")
+    self.btn_exit.setStyleSheet("color: #cc0000; font-weight: bold;")
+    self.btn_exit.clicked.connect(self.close)
+    self.layout.addWidget(self.btn_exit)
+    self.setLayout(self.layout)
